Split Sonarr models out into multiple files.
