+++ crates/webhook_api/src/models/sonarr/extended_episode_rating.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+/// A rating of a particular episode of a show.
+#[derive(Serialize, Deserialize, Debug, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SonarrExtendedEpisodeRating {
+    pub votes: u32,
+    pub value: serde_json::Number,
+}
